fix server crash due util import in integration file
